Add more composite workflows (#126)
Continuing to go down my existing workflows and update them to:
* Build out more bootstrapping workflows
* Test out having larger workflows, the new `cwd` step
* Get the templates more consistently organized

So changes include:
* Updating template folders to just be named "templates", putting all
workflows in the same package into that folder.
* Minor fixes to the `express/add-handler` workflow
* Add `express/init` workflow
* Add `openapi/init` workflow
* Add `openapi/add-schema` workflow
* Add `openapi/add-route` workflow
* Remove `openapi/update-schema` workflow
* Add a test workflow for `openapi` and `express` workflows together
* Fix "saf-specs generate" command to create a schemas folder if it DNE
* Fix "saf-docs generate-all" to *actually* update all the docs properly

A few changes to the workflows cli tool as well:
* Merged `kickoff` and `kickoff-unlisted`, realizing how to do so better
with commander.
* Fixed the status command.
* Added "skipIf" input property for command step
* Fixed "cwd" step to take absolute paths
* Add "valid" input property for update step
* Fixed my recursion logic for continuing a workflow machine
* A few documentation fixes
* Updates to the "add-workflow" command.

Regarding the "add-workflow" command, honestly I'm having a bit more
trouble with this workflow, in particular getting the agent to set up
template files correctly. This is maybe a little more advanced, unusual,
and important than other workflows to get right, so it's probably the
best one to test new methods of reliability.