Refactor link, router logic a bit (#128)
Reworking things to enable more than one product in a repo. This
requires letting "domain" be more than two segments, so I can have
"app1.docker.localhost" and "app2.docker.localhost" both serve as
domains for local development.

*That* means reworking the links and vue libraries a bit, moving some of
the vue-specific link logic to the vue library, and updating the "client
name" to always be the subdomain. This way the client name is consistent
(no need for it to be different than the subdomain) and also the link
logic can use the subdomain to determine what the "domain" is.

For this to work, I also need logic which requires the subdomain to be
set (like the sdk client) to run after the frontend code even has a
chance to set the client name/subdomain.

In summary, figuring out how to have two separate products live in the
same repo and deployment made me have to adjust and fix some things.

Also Cursor has started complaining about the typescript server pushing
against its memory limit... I might need to figure out how to optimize
types. I threw some quick adjustments in this change to see if it helps,
but I'll need a proper set up and experimentation to really make headway
there.