Add service/init workflow (#130)
Experimenting more with longer form workflows.

In this case, I'm generating all the typical constituent packages of a
service:
* `service-db`
* `service-http`
* `service-sdk`
* `service-spec`
* `service-common`
* `service-service`

This was mainly an exercise in templates, and how to structure workflows
in a way that they fit together. I worry a bit about how brittle this
structure is, but on the other hand, I've got a good web of type safety
and the potential for unit testing that I think I'll be okay.

It turns out also this workflow has no agentic components. It's just
copy/update. That's more reliable that way, and I don't have to worry
about code getting out of sync because if the templates are typed
correctly, then the outputs are likely to be as well as long as I don't
change the string replace logic.

Also:
* Added a "run-scripts" command to run a workflow without stopping at
any prompts or running any validation checks. This was useful so that I
could rapidly iterate on the templates and copy logic, though it turns
out I didn't need it because the workflows didn't have any prompts.
Still, I'll use this for more prompt-heavy workflows to debug.
* Updated the internal "dryRun" boolean to an enum so I can have more
run modes. One mode I expect to add is a "execute" mode where it calls a
CLI agent coding tool itself, rather than printing out prompts.
* Tweaked some names, general improvements to the existing init
workflows
* Standardized the init workflows on how the rename things, remove
prompting, get rid of steps that are "uncomment thing, fix it" which
were agentic bandaids.
* Small fix for monorepo/add-package
* Add sdk/init workflow to include in service/init workflow