Add grpc workflows (#132)
Working toward moving my secrets to a safer location, and separating
them from env config. But as part of that I need a grpc endpoint or two,
and so I figured I'd create some grpc workflows in prep.

I think perhaps this was the wrong choice.

It's tough to build a workflow outside of the context of actually
*using* it, and double so for grpc which is something I haven't used
nearly as much as other libraries in this framework. So trying to reason
about and come up with these workflows was harder than it needed to be.

They're also not quite the same structure as other libraries, and that
structure is awkward to template.

I'm going to merge what I have, and then iterate on these workflows as I
go, as well as add new one, more like I have in the past. I'd *like* to
make all the workflows ahead of time but I don't think that works so
well.

I also tried adding git commits along the way in the workflow, but it
ends up making things less clear in the process. I think what I'd rather
do is have the workflow *stop* at key points for review and then
fix/commit changes myself.

Also experimented with not git-ignoring the workflow status files. I
think I may still do that, but in conjunction with the above.

I'm keen on having the templates all play nice with one another, but I
think that's going to require each "init" having at least one "thing" in
it, so initializing a database needs one query, a server one handler, a
spec one route, etc.

Anyway, I added some notes to a stub doc for writing workflows. Got some
learnings from all of this at least.

Workflows added:
- `proto/add-method`
- `proto/init`
- `grpc/add-handler`
- `grpc/init`

Also:
* I added the ability to copy directories, not just files, as templates.
However, I didn't bother to rename the template placeholders in those
dirs... I may yet do that, still experimenting with this feature.
* Some fixes to workflows/add-workflow
* Added a spec and an aborted workflow-adding workflow for the secret
store. I'm not going to try and one-shot this thing as of yet... but
I'll break it down into larger-than-previously-done workflows.