Add secret store spec (#135)
Continuing kicking the tires on composite workflows! I like it.

Steps I took:
* Prompted the agent to make a plan similar to the last one.
* Needed a fair amount of rework, gave feedback, fixed
* Sent the agent on its merry way
* It reported at the end that some had been implemented, but some had
not due to file name conflicts. I realized the workflow was generating
files the wrong way, even though I'd checked. Not close enough
apparently.
* Fixed the workflows
* Had the agent run through the whole workflow again, encouraging reuse
of what they did.

So, let that be a lesson, I was doing the right thing (running the
workflow with run-scripts first) but not doing it right (I could have
noticed the generated files were in the wrong place before I sent the
agent along).