Implement @saflib/secrets-sdk tanstack queries (#137)
Making my way through.

Mainly I just added an `sdk/add-query` workflow and used it to generate
all the tanstack queries and mutations for each route spec.

For the new workflow, I'm trying something new: more visible
interpolation. I'm trying marking all code-to-replace with double
underscores. This is still valid JS/TS for names (unlike curly braces
which the agent would try to do by default), so the code can still run,
be type checked, etc, and it's a clear visual indication where in the
file things will be replaced.

The other thing I'm learning is because the templates all sort of
interact with each other, I need those names to be consistent across
workflows. I've previously just used "templateFile" for everything, but
if a file is used as a template in one workflow and is referenced *by* a
template file in *another* workflow, then this system breaks down.

Anyway, I'm seeing opportunity to improve the "copy" step interface, and
to share interpolation strings across workflows. This would very much
simplify/replace the context and line-replace code I've been writing.

Also:
* Handled an error where there's no checklist after running a workflow
* Some reformatting
* Fixed some specs that has the API provide the user's email. In
re-perusing the specs and tests this is the one thing that stuck out as
sub-optimal for how this backend system was built and designed.
* Added query params to the typed-msw implementation, so handlers get
them and can be reflected in the response. This was mainly because in
testing, I found the initial implementation wasn't providing query
params the right way to the openapi-fetch client. So something I want
these unit tests to do is make sure query params are properly provided.
* Fixed sdk to have type checking script