Improve workflow templating (#138)
Well, I finally got tired of the gnarly templating I was doing in
template files, where each workflow had its own unique set of strings to
look for and had no indication what strings were to be swapped out, and
no checks. So I built a few helper functions for workflows that use the
copy step, to parse out common values from paths and package names, and
to swap them in.

There are a bunch of conventions this depends on, so in a way it's
brittle, but it also enforces those conventions in a way that I think
will be worth it for keeping the overall codebase similarly patterned.

* Package names are structured `<service>-<package-type>`, like
"identity-http", or "secrets-db".
* All packages which depend on a particular package share that same
suffix (all -http packages depend on, and are generated from workflows
in, `@saflib/express`)
* All packages in the same service are part of the same organization
(the "@" prefix in the package name)
* For a "target", like a route or handler or query or page, they are
often organized into "groups", like route resources, database tables,
etc. And the files for these targets are organized into folders named as
those groups.

From these conventions, the templating system comes with a bunch of
"free" values if you provide them some inputs:
* Package name -> service-name, organization-name, shared-package-prefix
* Target path -> group-name, target-name

And, names need at least two parts *so that* provided names are
automatically converted from kebab-case to camelCase, PascalCase,
snake_case, and SNAKE_CASE. So now all my templates have things like
`__service-name__` or `__targetName__` interspersed which provides a
consistent, package-scoped set of names to work with, and with these
helpers and the workflows themselves having consistent inputs, it's a
whole lot more manageable than before.

I didn't update all the workflows, but I did it for the ones where
things were getting pretty wobbly: the related workflows for creating a
service: drizzle, express, service, openapi, etc.

Oh yeah, there was *one* concession I had to make to the templating:
package names can't start with "__" so the template packages have
`template-package-...` for names and those are also replaced
specifically, but it's less clear that they are.