Implement @saflib/secrets-grpc-* (#142)
This ended up being quite involved.

I reworked and added all the grpc workflows as part of this PR. There
are six of them, two for each of making a proto library, a client
library, and a server library. One to initialize, the other to add to
it. I also merged the "grpc" and "grpc-spec" packages into a single one,
figuring there's less need to separate dependencies in this case than
for frontend.

It was also a fair amount of work just because proto's quite involved.
My approach with my packages here is really to contain the complexity,
and enable more features in the future. Also to make the way I've got
proto packages set up similar to how I've got http packages set up: with
client/server/spec separated and each of those sharing similar folder
structures based on http-resource/grpc-service. And the whole fake
implementation.

I was a bit disappointed with how the agent tried to implement the
server handlers, though. I basically had to write those myself, but
that's for the best. Still! might want to think about how to get agents
to build something closer to what is intended. I think maybe some
pseudocode would be good, especially if it highlights what database
queries are needed. I ended up having to go back and make some changes
to the database, and I think that's just a reflection of how I should
have thought things through more going in.

Also:
* Fixed a small issue with the drizzle add-query workflow.
* Added support for template files in nested folders, by keeping their
folder structure.
* Added a check, if a line says "DELETE_THIS_LINE" in it... it's deleted