Update workflows, add backup and download functionality, add back package-lock.json (#162)
* Revert back to having a package-lock.json file. As expected, it was
causing some instability in CI.
* Rework cron a bit, have it return created jobs and run the async
storage.
* Add cron/init workflow and cron/add-job workflow, along with template
files. Note: I think in the future I'll incorporate cron into
service/init, because I'd like backups to be a standard feature and cron
will drive this.
* Add backup interface to drizzle instances
* Share logging a bit better in node errors.ts. Note: I need to rethink
how logging is done a bit. It seems a bit confusing, maybe overly
complicated.
* Refactor saflib/sdk into a src folder.
* Add download helper to saflib/sdk
* Fix sdk/add-query workflow - for some reason it was prompting to have
both a mutation and a query
* Modernize and simplify workflows/add-workflow. However, it's still not
great... this might be one of the harder workflows to get right, since
it's pretty outside of the norm.