Add workflow integration tests, clean up workflows (#170)
So many commits!

I've made a great many investments in the workflow tool which I'd like
to publish, so I'm adding tests to ensure the tool and my framework's
workflows (which serve as examples) are reliable. This PR is me adding a
bunch of tests and fixing a bunch of things those tests uncovered. All
the tests are added as github actions.

## Version Control
I'm finding using version control to enforce editing the right files by
the agent to be very useful, and so I've made it now the default
behavior and improved the current interface. Changes:

* Rename `manageGit` to `versionControl`. Your configuration here is not
strictly tied to git, and you no longer opt into git management but
instead configure how you want version control to work. Right now it
just has an updated "allowPaths" field which takes minimatch globs. Use
these if the workflow generates or edits files other than the template
files.
* Refactor the git logic into its own separate file.
* Update all workflows to include necessary allowed paths, and added a
`test-git.sh` file which runs in CI to ensure every workflow
successfully finishes when there are no agentic changes.

## Print/Run/Script Mode
You can run workflows in various ways, where the tool *prints* prompts
to the agent, the tool *runs* the agent passing in the prompt, or the
tool skips past all prompting and just runs the *scripts*. I added a
github action for each one.

The script mode executes a workflow which includes every other workflow.
The other two I didn't think needed that same level of rigor done again
and again so I just have them run a simple workflow which includes one
of every non-trivial step.

## Foundational Improvements
Besides fixing a host of workflow-specific issues, and all that git
stuff, I also used this time to make the core logic work better and add
some features, including:

### CLI stuff
* A version control argument to opt-in to having workflows make git
commits, and enforce allowed paths through that.
* A skip todos argument to allow the tool to not get stopped when TODO
strings are still in files as part of update steps.
* "mock" as an option for run mode, so that that mode can be tested.
* Add "all" and "details" options to the "list" command. Getting a list
of all workflows was previously handled by running `help kickoff` since
each workflow was an explicit subcommand but that's been broken for a
while and wasn't really the best place anyway.
* Sort these listed workflows alphabetically
* "run-scripts" command exits with code 1 when the workflow fails to
complete.
* Put an FSM state dump into a separate file when it ends in an error,
for debugging. Don't save the state to the normal location when any of
the actors are "in error" so that the workflow might not be in an
irrecoverable state.
* Only print workflow duration in run mode.
* Don't create a log file in anything but run mode. It's the only mode
that uses it.
* Add logging in script mode to make it easier to see what step we're
on.

### Core stuff
* Removed `afterEach`, which I only ever used to do some hacky git
commit stuff.
* Make steps which prompt the agent work consistently in terms of when
you run `npm exec saf-workflow status` it printing the current prompt,
and it continuing when you run `npm exec saf-workflow next`.
* Add some error logging in the copy step. Better errors generally.

Essentially, adding the tests got me exactly what I wanted: the tool and
the workflows work much more reliably, and I feel confident they will
continue to do so as I continue to work on them.

## Future Work

There's definitely more I can do in terms of testing. The sorts of
things that are still todo:
* Better unit testing. I think honestly part of that will be pulling
some of the step logic out of xstate entirely into a place where I can
better wrangle it. I might try having all steps be mainly a promise in a
common machine shell.
* Evals. Make sure these workflows are designed well such that an agent
can consistently complete them in a variety of scenarios.
* Jest integration testing. It would be great to know what my coverage
is for everything I'm doing, rather than having everything be a custom
script. I may at some point convert things over. As it is, I'm *pretty*
sure I've covered the most important code paths.

Okay that's enough.