Add product/init workflow (#180)
I'm about to start my third product, and at this point I've pretty well
defined how I want my stack set up, I've got a cadre of routine
workflows, and I'm getting a fair amount done with complex workflows.
Now I'm seeing how much I can use the workflow tool to get a whole new
product stood up with all the trimmings that I think should come with
any non-trivial or marketable product.

At the end of this, #169 and #154 should be done. They're not quite, but
this PR is a large down payment.

The main driver is the new product/init workflow, but to support that,
this PR also includes:
* Add @saflib/backup, a series of packages which extend a service to
backup its database on the regular.
* Extend @saflib/drizzle to have the database provide a "restore"
function.
* Add saflib/object-store, which provides a generic interface for
working with a blob store, and the azure implementation
* Add @saflib/sentry package
* Update @saflib/account-sdk pages and router function to allow more
fine-grained control of redirects

Some fixes along the way
* Modernize monorepo/add-export workflow
* Don't have the loki transport error if loki env not provided
* Some improvements to processes/spec-project workflow
* Bunch of vue/add-spa improvements
* Some workflow tool fixes, like fixing how organization is parsed in
templating, continuing the workflow recursively properly, and fixing a
runtime error with the cursor agent when read tool fails