Add vue/add-e2e-test workflow (#182)
A cornerstone of getting an agent to build something correctly end to
end is them being able to add end-to-end tests. For that, I need a
rock-solid workflow to add them, and strong support in the framework to
make adding new tests easy, and making those tests durable.

It's tough.

This work gets me *part* of the way there. I iterated a few times on a
playwright flow, improving the framework to reduce errors on the agent's
part, but it requires writing good fixtures and helper methods. I'm only
starting to play around with playwright fixtures so I still probably
have more to learn there, and I made some improvements to my existing
helpers. Aside from adding the vue/add-e2e-test workflow and extending
vue/add-spa and vue/add-page to include fixtures, this PR also does the
following:

* **Add saflib/admin package** - My playwright tests depend on there
being an admin spa with some helpful tools, like getting sent emails and
entering "test mode" where events are exposed and vue dev tools are
hidden.
* **Fix dependency collection to not mind so much circular deps in
packages.** - I forget exactly what did this, but it's valid for two
packages to import one another, as long as they don't import exports
that import each other.
* **Add fixtures to saflib/auth** - to help with a playwright test
registering, logging in, etc
* **Improve option select in playwright helper** - it's actually pretty
tough to get v-select to work in a playwright test, mainly I think
because not every option is rendered.